% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_fair_theory.R
\name{add_theory_file}
\alias{add_theory_file}
\title{Add Theory File}
\usage{
add_theory_file(path, theory_file = "theory.txt")
}
\arguments{
\item{path}{Character, indicating the directory in which to create the FAIR
theory.}

\item{theory_file}{Character, referring to existing theory file(s) to be
copied, or a new theory file to be created. Default \code{NULL} does nothing.}
}
\value{
Invisibly returns a logical value,
indicating whether the function was successful or not.
}
\description{
Writes a theory file to a specific path.
}
\examples{
add_theory_file(path = tempdir(), theory_file = "theory.txt")
}
