% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_pkgdown.R
\name{add_to_pkgdown}
\alias{add_to_pkgdown}
\title{Add webexercises helper files to pkgdown}
\usage{
add_to_pkgdown(pkgdown_dir = ".")
}
\arguments{
\item{pkgdown_dir}{The base directory for your pkgdown project}
}
\value{
No return value, called for side effects.
}
\description{
Adds the necessary helper files to an existing pkgdown project.
}
