% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{cosine_similarity}
\alias{cosine_similarity}
\title{Cosine Similarity}
\usage{
cosine_similarity(x)
}
\arguments{
\item{x}{Numeric matrix, where rows are cases and columns are
features.}
}
\value{
Matrix.
}
\description{
Compute cosine similarity between rows of matrix.
}
\examples{
set.seed(1)
cosine_similarity(matrix(runif(30), nrow = 3))
}
