% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{download_huggingface}
\alias{download_huggingface}
\title{Download 'Hugging Face' Model}
\usage{
download_huggingface(model, path, ...)
}
\arguments{
\item{model}{Atomic character, referencing a model - typically of the form
\code{"user/model"}.}

\item{path}{Atomic character, referencing a local folder where the model is
installed.}

\item{...}{Arguments passed to and from functions.}
}
\value{
On success: Atomic character, with the path to the local model.
On failure: \code{NULL}.
}
\description{
Download a model from 'Hugging Face' to a local folder, to be
used in e.g. \link[theorytools]{get_embeddings}.
}
