% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_theory.R
\name{download_theory}
\alias{download_theory}
\title{Download FAIR Theory}
\usage{
download_theory(id, path = ".")
}
\arguments{
\item{id}{URL of the 'Git' repository or DOI of the 'Zenodo' archive.}

\item{path}{Character, indicating the directory in which to create the FAIR
theory.}
}
\description{
Downloads a FAIR theory archive from a 'Git' remote repository or 'Zenodo'.
}
\examples{
download_theory(id = "https://github.com/cjvanlissa/tripartite_model.git",
path = file.path(tempdir(), "tripartite_git"))
download_theory(id = "10.5281/zenodo.14921521",
path = file.path(tempdir(), "tripartite_zenodo"))
}
