% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenodo.R
\name{download_zenodo}
\alias{download_zenodo}
\title{Download Zenodo Archive}
\usage{
download_zenodo(doi, path = ".")
}
\arguments{
\item{doi}{DOI of the Zenodo archive.}

\item{path}{Character, indicating the directory in which to create the FAIR
theory.}
}
\description{
Downloads an archive from \href{https://zenodo.org}{Zenodo}.
}
\examples{
download_zenodo(doi = "10.5281/zenodo.14921521",
path = file.path(tempdir(), "tripartite"))
}
