% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{similarity_cosine}
\alias{similarity_cosine}
\title{Cosine Similarity}
\usage{
similarity_cosine(x)
}
\arguments{
\item{x}{Numeric matrix, where rows are cases and columns are
features.}
}
\value{
\code{matrix}
}
\description{
Compute cosine similarity between rows of matrix.
}
\examples{
set.seed(1)
similarity_cosine(matrix(runif(30), nrow = 3))
}
