% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{similarity_stringdist}
\alias{similarity_stringdist}
\title{String Similarity}
\usage{
similarity_stringdist(x, ...)
}
\arguments{
\item{x}{Character vector.}

\item{...}{Arguments passed to \link[stringdist]{stringdistmatrix}.}
}
\value{
\code{matrix}
}
\description{
Wraps \link[stringdist]{stringdistmatrix} to return a string
similarity matrix (e.g., for use with \link[theorytools]{code}).
}
\examples{
similarity_stringdist(c("cat", "a cat", "dog"))
}
