% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_pkgdown.R
\name{use_webex_vignette}
\alias{use_webex_vignette}
\title{Create a webexercises vignette}
\usage{
use_webex_vignette(name, title = NULL, type = c("vignette", "article"))
}
\arguments{
\item{name}{Atomic character, vignette name. See
\link[usethis]{use_vignette}.}

\item{title}{Atomic character, vignette title. See
\link[usethis]{use_vignette}.}

\item{type}{Atomic character, one of \code{c("vignette", "article")}, defaults to
\code{"vignette"}.}
}
\value{
Returns \code{NULL} invisibly, called for its side effects.
}
\description{
Wraps \link[usethis]{use_vignette} to add a vignette or article to
\verb{vignettes/} with support for \code{webexercises}.
}
\examples{
\dontrun{
use_webex_vignette("vignette_with_quiz.Rmd", "Quiz people with webexercises")
}
}
