% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webex_vignette.R
\name{webex_vignette}
\alias{webex_vignette}
\title{Create Vignette with webexercises Support}
\usage{
webex_vignette(...)
}
\arguments{
\item{...}{Additional function arguments to pass to
\code{\link[rmarkdown]{html_document}}.}
}
\value{
R Markdown output format to pass to 'render'.
}
\description{
This function wraps \code{rmarkdown::html_document} to configure
compilation to embed the default webexercises CSS and JavaScript files in
the resulting HTML.
}
\details{
Call this function as the \code{output_format} argument
for the \code{\link[rmarkdown]{render}} function when compiling
HTML documents from RMarkdown source.
}
\seealso{
\code{\link[rmarkdown]{render}}, \code{\link[rmarkdown]{html_document}}
}
