% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_fair_theory.R
\name{add_readme_fair_theory}
\alias{add_readme_fair_theory}
\title{Add Readme File}
\usage{
add_readme_fair_theory(path, title, ...)
}
\arguments{
\item{path}{Character, indicating the directory in which to create the FAIR
theory.}

\item{title}{Character, indicating the theory title. Default: \code{NULL}}

\item{...}{Additional arguments passed to other functions.}
}
\value{
Invisibly returns a logical value,
indicating whether the function was successful or not.
}
\description{
Writes a README file to a specific path.
}
\examples{
add_readme_fair_theory(path = tempdir(), title = "My Theory")
}
